ingestion-pipeline/requirements.txt
pandas==2.2.2
numpy==1.26.4
sentence-transformers==3.0.0
qdrant-client==1.9.1
pyyaml==6.0.1
loguru==0.7.2
tqdm==4.66.4